requirements.txt
flask
gunicorn
requests
groq
python-dotenv
beautifulsoup4
PyJWT